Move profile, certificate, and utility functions into corresponding go files
